requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==3.2.15
django-cors-headers==3.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
uuid==1.30
